Car detail current
made this while practicing C++. Inspiration was my GF's (soon to-be fiance) job where she worked at a car detail shop. they had stupid basic and very terrible systems there and this would have been a great help to them. ther will be more added to this later but this is a basic start with what i know at the moment to keep me studying and show my work in case anyone wants to see.